// bank/src/types/db.d.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      inserted_checks: {
        Row: {
          amount: number
          check_date: string
          check_name: string
          id: string
          inserted_at: string | null
        }
        Insert: {
          amount: number
          check_date: string
          check_name: string
          id: string
          inserted_at?: string | null
        }
        Update: {
          amount?: number
          check_date?: string
          check_name?: string
          id?: string
          inserted_at?: string | null
        }
        Relationships: []
      }
// ...
      transactions: {
        Row: {
          account_id: number
          amount: number
          balance: number
          created_at: string
          description: string
          id: number
          status: Database["public"]["Enums"]["transaction_status_enum"]
          transaction_type: Database["public"]["Enums"]["transaction_type_enum"]
        }
        Insert: {
          account_id: number
          amount: number
          balance: number
          created_at?: string
          description: string
          id?: number
          status?: Database["public"]["Enums"]["transaction_status_enum"]
          transaction_type: Database["public"]["Enums"]["transaction_type_enum"]
        }
        Update: {
          account_id?: number
          amount?: number
          balance?: number
          created_at?: string
          description?: string
          id?: number
          status?: Database["public"]["Enums"]["transaction_status_enum"]
          transaction_type?: Database["public"]["Enums"]["transaction_type_enum"]
        }
        Relationships: [
          {
            foreignKeyName: "transactions_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never